requirements.txt
anthropic>=0.40.0
elevenlabs>=1.0.0
google-cloud-texttospeech>=2.16.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
